refactor/docs: Added some KDoc and refactored where it was necessary. The code now follows the structure of the application
