+++ app/src/main/java/no/usn/mob3000/data/source/local/content/NewsDao.kt
@@ -1,4 +1,4 @@
-package no.usn.mob3000.data.network.local
+package no.usn.mob3000.data.source.local.content
 
 import androidx.room.Dao
 import androidx.room.Delete
@@ -6,7 +6,14 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
+import no.usn.mob3000.data.model.content.local.NewsItemLocal
 
+/**
+ * Data Access Object (DAO) interface for the news table.
+ *
+ * @author 258030
+ * @created 2024-11-13
+ */
 @Dao
 interface NewsDao {
