+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/content/DocumentationViewModel.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
-import no.usn.mob3000.data.repository.content.DocsRepository
+import no.usn.mob3000.data.repository.content.remote.DocsRepository
 import no.usn.mob3000.domain.model.content.DocsData
 import no.usn.mob3000.domain.usecase.content.docs.DeleteDocsUseCase
 import no.usn.mob3000.domain.usecase.content.docs.FetchDocUseCase
